+++ server/server.js
@@ -9,18 +9,12 @@ import socketIO from "socket.io";
 import http from "http";
 import ChatRouter from "./Router/chats.js";
 import ProfileRouter from "./Router/profile.js";
-import { tokenCheck } from "./token.js";
 import cookieParser from "cookie-parser";
+import socketServer from "./socket.js";
 
 const app = express();
 export const server = http.createServer(app);
 app.use(cookieParser());
-const io = socketIO(server, {
-  cors: {
-    origin: "*",
-    methods: ["GET", "POST"],
-  },
-});
 
 connectMongoDB();
 
@@ -48,25 +42,7 @@ app.use("/api/profile", ProfileRouter);
 
 const port = 8080; // Node 서버가 사용할 포트 번호
 
-// io.of() 채널 만들어주는 메서드, "/chat 채널"
-const chat = io.of("/chat").on("connection", (socket) => {
-  console.log("Socket connected!");
-
-  socket.on("enter_room", ({ roomId }) => {
-    // "socket join 메서드를 사용하면 인자로 전달한 방으로 연결."
-    socket.join(roomId);
-  });
-  // 클라이언트에서 연결 해제 이벤트를 처리
-  socket.on("disconnect", () => {
-    console.log("Socket disconnected!");
-  });
-
-  socket.on("message", (message, { roomId }) => {
-    //해당 채팅방으로 메시지를 보낸다.
-    chat.to(roomId).emit("message", message);
-    // io.emit("message", { name, message });
-  });
-});
+socketServer(server);
 
 server.listen(port, () => {
   console.log(`Listening on port ${port}`);
